reorganise access to mock data
